Add a convenient mechanism for observing media status (#2)
This seems like a common enough use-case that it seems to make sense to
have a high-level primitive for it, since doing so by hand is a bit...
tricky. Basically, any time the active app changes (indicated by a
`RECEIVER_STATUS` message) we will need to re-subscribe to the new app's
media channel in order to get updates from it.